# swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/v1/purchase":{"post":{"operationId":"PurchaseController_purchaes","summary":"","description":"This API performs, to make purchase.","parameters":[{"name":"customerId","required":true,"in":"query","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePurchaseDto"}}}},"responses":{"200":{"description":"Success.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseCreatePurchaseDto"}}}}},"tags":["Purchase"]}},"/api/v1/purchase/customer":{"post":{"operationId":"PurchaseController_createCustomer","summary":"","description":"This API performs, Process a user purchasing a dish from a restaurant, handling all relevant data changes in an atomic transaction.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerDto"}}}},"responses":{"201":{"description":"Initiate sell for a customer.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseCreateCustomerDto"}}}}},"tags":["Purchase"]}},"/api/v1/purchase/customer/{id}":{"patch":{"operationId":"PurchaseController_customerStatusUpdate","summary":"","description":"This API performs, to update customer information.","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerDto"}}}},"responses":{"200":{"description":"Update a customer.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerDto"}}}}},"tags":["Purchase"]}},"/api/v1/purchase/customers":{"get":{"operationId":"PurchaseController_findAllCustomers","summary":"","description":"This API performs, to update customer information.","parameters":[{"name":"q","required":false,"in":"query","example":"search string","schema":{"type":"string"}},{"name":"sort","required":false,"in":"query","example":"object.key|columnName","description":"use \"-\" at the beginning to sort by descending order","schema":{"type":"string"}},{"name":"offset","required":false,"in":"query","example":0,"schema":{"type":"number"}},{"name":"limit","required":false,"in":"query","example":10,"schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerDto"}}}},"responses":{"200":{"description":"Get customer list.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseCustomersDto"}}}}},"tags":["Purchase"]}},"/api/v1/purchase/invoice":{"get":{"operationId":"PurchaseController_findInvoice","summary":"","description":"This API performs, to return invoice.","parameters":[{"name":"q","required":false,"in":"query","example":"search string","schema":{"type":"string"}},{"name":"sort","required":false,"in":"query","example":"object.key|columnName","description":"use \"-\" at the beginning to sort by descending order","schema":{"type":"string"}},{"name":"offset","required":false,"in":"query","example":0,"schema":{"type":"number"}},{"name":"limit","required":false,"in":"query","example":10,"schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerDto"}}}},"responses":{"200":{"description":"Get invoice.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseInvoicesDto"}}}}},"tags":["Purchase"]}},"/api/v1/purchase/customers/{id}":{"get":{"operationId":"PurchaseController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseACustomerDto"}}}}},"tags":["Purchase"]}},"/api/v1/products":{"post":{"operationId":"ProductController_create","summary":"","description":"This API performs to create product","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductDto"}}}},"responses":{"201":{"description":"Successfully created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseCreateProductDto"}}}}},"tags":["products"]},"get":{"operationId":"ProductController_findAll","parameters":[{"name":"q","required":false,"in":"query","example":"search string","schema":{"type":"string"}},{"name":"sort","required":false,"in":"query","example":"object.key|columnName","description":"use \"-\" at the beginning to sort by descending order","schema":{"type":"string"}},{"name":"offset","required":false,"in":"query","example":0,"schema":{"type":"number"}},{"name":"limit","required":false,"in":"query","example":10,"schema":{"type":"number"}}],"responses":{"200":{"description":"get the list of all records","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseProductsDto"}}}}},"tags":["products"]}},"/api/v1/products/inventory":{"post":{"operationId":"ProductController_createInventory","summary":"","description":"This API performs to create inventory of a product","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductDto"}}}},"responses":{"201":{"description":"Successfully created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseCreateProductDto"}}}}},"tags":["products"]}},"/api/v1/products/{id}":{"get":{"operationId":"ProductController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseAProductDto"}}}}},"tags":["products"]},"patch":{"operationId":"ProductController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductDto"}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductDto"}}}}},"tags":["products"]},"delete":{"operationId":"ProductController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["products"]}},"/api/v1/products/inventory/{id}":{"patch":{"operationId":"ProductController_updateInventory","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateInventoryDto"}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseUpdateInventoryDto"}}}}},"tags":["products"]}},"/api/inventory":{"post":{"operationId":"InventoryController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateInventoryDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"InventoryController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/api/inventory/{id}":{"get":{"operationId":"InventoryController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"InventoryController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateInventoryDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"InventoryController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}}},"info":{"title":"shop-APIs","description":"","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"RequestBodyProductDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid","description":"Product table primary key"},"qty":{"type":"number","description":"Quantity of product"}},"required":["id","qty"]},"CreatePurchaseDto":{"type":"object","properties":{"productList":{"type":"array","items":{"$ref":"#/components/schemas/RequestBodyProductDto"}},"status":{"type":"string","enum":["completed","return","cancelled"],"default":"completed"},"discountInPercentage":{"type":"number","description":"Discount in percentage."},"vatInPercentage":{"type":"number","description":"VAT in percentage."},"createdBy":{"type":"string","format":"uuid","description":"Created by."}},"required":["productList","status","discountInPercentage","vatInPercentage","createdBy"]},"ProductDetailDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid","description":"Purchase primery key."},"purchaseCode":{"type":"string","description":"Unique code of particuler purchase"},"customerId":{"type":"string","format":"uuid","description":"Customer Id."},"discount":{"type":"number","description":"Discount amount"},"total":{"type":"number","description":"Total amount"},"vat":{"type":"number","description":"VAT amount"},"payable":{"type":"number","description":"Payable"},"status":{"type":"string","enum":["completed","return","cancelled"],"default":"completed"},"created":{"format":"date-time","type":"string","example":"2022-08-14T12:21:27.612Z","description":"Date of purchase"},"updated":{"format":"date-time","type":"string","example":"2022-08-14T12:21:27.612Z","description":"Date of update purchase"},"createdBy":{"type":"string","format":"uuid","description":"Created by."},"updatedBy":{"type":"string","format":"uuid","description":"Created by."}},"required":["id","purchaseCode","customerId","discount","total","vat","payable","status","created","updated","createdBy","updatedBy"]},"SoldProductDetailDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid","description":"Product code"},"productId":{"type":"string","format":"uuid","description":"Product table primary key"},"purchaseId":{"type":"string","format":"uuid","description":"Purchase table primary key"},"qty":{"type":"number","description":"Quantity of the product"},"unitPrice":{"type":"number","description":"Unit price of the product"},"created":{"format":"date-time","type":"string","example":"2022-08-14T12:21:27.612Z","description":"Date of purchase"},"updated":{"format":"date-time","type":"string","example":"2022-08-14T12:21:27.612Z","description":"Purchase data update date"},"deletedAt":{"format":"date-time","type":"string","example":"2022-08-14T12:21:27.612Z","description":"Purchase data delete date"}},"required":["id","productId","purchaseId","qty","unitPrice","created","updated","deletedAt"]},"ResponseCreatePurchaseDto":{"type":"object","properties":{"detail":{"$ref":"#/components/schemas/ProductDetailDto"},"products":{"type":"array","items":{"$ref":"#/components/schemas/SoldProductDetailDto"}}},"required":["detail","products"]},"CustomerDto":{"type":"object","properties":{"name":{"type":"string","example":"Mr. xyz","description":"User name/ Customer name!"},"address":{"type":"string","example":"Customer address","description":"House#20, Road#30, Block#C, Dhaka"},"contactNumber":{"type":"string","example":"8801999999999","description":"Customer contact number"},"type":{"type":"string","enum":["old","new"],"default":"old"},"createdBy":{"type":"string","description":"Created by."}},"required":["name","address","contactNumber","type","createdBy"]},"ResponseCreateCustomerDto":{"type":"object","properties":{}},"PaginationDto":{"type":"object","properties":{"term":{"type":"string","example":"search string","readOnly":true},"sort":{"type":"string","example":"object.key | ColumnNName","description":"the column name or object key based on which sorting is done","readOnly":true},"offset":{"type":"number","example":10},"limit":{"type":"number","example":10},"previousOffset":{"type":"number","example":0,"nullable":true},"nextOffset":{"type":"number","example":20,"nullable":true},"currentPage":{"type":"number","example":1},"pageCount":{"type":"number","example":5},"total":{"type":"number","example":50}},"required":["offset","limit","previousOffset","nextOffset","currentPage","pageCount","total"]},"ResponseCustomersDto":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/CustomerDto"}},"metadata":{"$ref":"#/components/schemas/PaginationDto"}},"required":["items","metadata"]},"ProductDto":{"type":"object","properties":{"PRODUCT_CODE":{"type":"string","example":"bl543jr","description":"Product code"},"NAME":{"type":"string","description":"Customer name"},"BRAND":{"type":"string","description":"Brand name"},"PRICE":{"type":"number","description":"Price of the product"},"QUANTITY":{"type":"number","description":"Purchase quantity"}},"required":["PRODUCT_CODE","NAME","BRAND","PRICE","QUANTITY"]},"ResponseInvoicesDto":{"type":"object","properties":{"DATE":{"format":"date-time","type":"string","description":"Date of transection"},"INVOICE_ID":{"type":"string","example":"QAb3ht3","description":"PurchaseCode/Invoice ID"},"CUSTOMER_NAME":{"type":"string","description":"Name of the customer"},"CUSTOMER_CONTACT":{"type":"number","example":"+99882323232","description":"Contact number of customer"},"CUSTOMER_ADDRESS":{"type":"string","description":"Address of the customer"},"CUSTOMER_TYPE":{"type":"string","enum":["old","new"],"description":"Contact number of customer"},"PRODUCTS":{"type":"array","items":{"$ref":"#/components/schemas/ProductDto"}},"TOTAL":{"type":"number","example":1000,"description":"Total amount"},"DISCOUNT":{"type":"number","example":100,"description":"Applicable discount amount"},"SUB_TOTAL":{"type":"number","description":"Sub total"},"VAT":{"type":"number","description":"Applicable VAT amount"},"PAYABLE":{"type":"number","description":"Payable amount"},"STATUS":{"type":"string","enum":["completed","return","cancelled"],"description":"Status of the purchase"}},"required":["DATE","INVOICE_ID","CUSTOMER_NAME","CUSTOMER_CONTACT","CUSTOMER_ADDRESS","CUSTOMER_TYPE","PRODUCTS","TOTAL","DISCOUNT","SUB_TOTAL","VAT","PAYABLE","STATUS"]},"ResponseACustomerDto":{"type":"object","properties":{}},"CreateProductDto":{"type":"object","properties":{"name":{"type":"string","description":"Product name."},"brandName":{"type":"string","description":"Brand name."},"price":{"type":"string","description":"Price."},"createdBy":{"type":"string","description":"Created by."},"codePrefix":{"type":"string","example":"AD","description":"Product prefix code."}},"required":["name","brandName","price","createdBy","codePrefix"]},"ResponseCreateProductDto":{"type":"object","properties":{"name":{"type":"string","description":"Product name."},"brandName":{"type":"string","description":"Brand name."},"code":{"type":"string","enum":["CD","AD","QA"],"default":"QA"},"price":{"type":"string","description":"Price."},"createdBy":{"type":"string","description":"Created by."}},"required":["name","brandName","code","price","createdBy"]},"ResponseProductsDto":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/ProductDto"}},"metadata":{"$ref":"#/components/schemas/PaginationDto"}},"required":["items","metadata"]},"ResponseAProductDto":{"type":"object","properties":{"name":{"type":"string","description":"Product name."},"brandName":{"type":"string","description":"Brand name."},"code":{"type":"string","enum":["CD","AD","QA"],"default":"QA"},"price":{"type":"string","description":"Price."},"createdBy":{"type":"string","description":"Created by."},"id":{"type":"string","description":"Unique identifier of a product."}},"required":["name","brandName","code","price","createdBy","id"]},"UpdateProductDto":{"type":"object","properties":{"name":{"type":"string","description":"Product name."},"brandName":{"type":"string","description":"Brand name."},"price":{"type":"string","description":"Price."},"inventoryStatus":{"type":"string","description":"Inventory status."}},"required":["name","brandName","price","inventoryStatus"]},"UpdateInventoryDto":{"type":"object","properties":{"productId":{"type":"uuid","description":"Product Id."},"totalQty":{"type":"string","description":"Total quantity."},"createdBy":{"type":"string","description":"Created by."}}},"ResponseUpdateInventoryDto":{"type":"object","properties":{"productId":{"type":"uuid","description":"Product Id."},"totalQty":{"type":"string","description":"Total quantity."}},"required":["productId","totalQty"]},"CreateInventoryDto":{"type":"object","properties":{"productId":{"type":"uuid","description":"Product Id."},"totalQty":{"type":"string","description":"Total quantity."},"createdBy":{"type":"string","description":"Created by."}},"required":["productId","totalQty","createdBy"]}}}}